[BREAKING] Use a symbol generator in term_fol
Implement chooseW and chooseWL which choose from a list based on a
weight. Fulfills a similar role to choose' which suffers from bad
performance and only works with integers. Additionally this replaces the
weird ifMultiple construct used before.
term_fol now uses a symbol generator which is passed in instead of 3
frequencies.
Use term_fol (def_sym_gen (w_bound, w_free, etc.)) for (almost)
identical behaviour.
Now a theoretically unlimited number of symbol generators can be used,
useful for e.g. starting with a
basic symbol generator and adding an additional one which represents a
recurring symbol whenever a new symbol is generated. That is, the
probability for a symbol being reused changes during generation:
0.0  (no symbol generated yet) -> 0.5 -> 0.67 -> 0.75 etc.